ai-service/requirements.txt
langgraph
langchain-google-genai
fastapi
uvicorn
pydantic
python-dotenv
